% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{bats}
\alias{bats}
\title{Movement data for Sturnira lilium bats}
\format{Three objects based on the same data: \code{bats}, a data frame with the raw
movement data; \code{bats_sf}, a spatial data frame of the class \code{\link{sf}};
and \code{bats_sp}, a \code{\link{SpatialPointsDataFrame}}.
These tables contain 767 rows and 7 variables:
\describe{
  \item{Tag_ID}{ID of radio collar}
  \item{Animal_ID}{ID of the individual bat}
  \item{x, y}{longitude and latitute of the record (UTM 23S, WGS84)}
  \item{date}{date of the record}
  \item{time}{time of the day of the record}
  \item{position}{whether the animal was in the roost or was foraging (vertical)
  or flying (horizontal)}
}}
\source{
What do we put here?
}
\usage{
data(bats)
data(bats_sf)
data(bats_sp)
}
\description{
Radiotelemetry movement data of \italic{S. lilium} frugivorous bats in the state of Sao Paulo.
}
\examples{
library(ggplot2)
library(sf)

# raw data table
data(bats)
bats

# plot
ggplot(bats) +
  geom_point(aes(x, y, color = Animal_ID)) +
  theme_minimal() +
  labs(color = "Animal_ID")

# sf object
bats_sf

# plot
ggplot(bats_sf) +
  geom_sf(aes(color = Animal_ID)) +
  theme_minimal()

# sp
head(bats_sp)

}
\references{
Kerches-Rogeri, P., Niebuhr, B.B., Muylaert, R.L, Mello, M.A.R.  Individual
specialization in the space use of frugivorous bats. Journal of Animal Ecology.
}
\keyword{data}
