% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatis.R
\name{spatis}
\alias{spatis}
\title{Calculation of individual specialization in the use of space from animal tracking}
\usage{
spatis(data, individuals.col, population.ID = NULL, index = c("spatis",
  "spatics"), method = c("VI", "HR", "PHR", "BA", "UDOI")[1], ...)
}
\arguments{
\item{data}{a SpatialPointsDataFrame object containing the locations of
the individuals, their IDs, and other spatial or individual
information.}

\item{individuals.col}{name (character) or number of the column of the input
SpatialPointsDataFrame 'data' that represents the identity of individuals.}

\item{population.ID}{ID (character or number) that represents the population in the
column 'individuals.col' of the input dataset 'data'. In case
the locations of all individuals were not pooled and combined
to the original data, population.ID is set to NULL (the default).
In this case, the calculations for the whole population are
made inside the SpatIS function.}

\item{index}{character or vector of characters indicating the indexes to be calculated
(\code{"spatis"} or \code{"spatics"}).}

\item{method}{method to calculate the overlap between utilization distributions
or areas of use (the same options as the function `adehabitatHR::kerneloverlap`,
except for the "HD" method; see ?adehabitatHR::kerneloverlap for more information).
The default method is "VI", which computes the volume of the intersection between
the individual and populational UDs.}

\item{...}{additional arguments to be passed to the function `adehabitatHR::kerneloverlap`
to calculate the overlap between UDs or to the function `adehabitat::kernelUD``
for the kernel estimation of the utilization distribution.}
}
\value{
The SpatIS function returns a list of six elements:
\item{data}{the data used to calculate SpatIS (with locations corresponding
    to the whole population, independently of whether the population
    locations were already in the input data). It is a SpatialPointsDataFrame.}
\item{parms}{a list with parameters used as input to call SpatIS function.}
\item{SpatIS.individual}{a vector of Spatial individual specialization indices for
    each individual of the population (the overlap between
    each individual and the population UDs).}
\item{SpatIS.population}{a value of Spatial individual specialization for the population,
    calculated as the average of all SpatIS individual values.}
\item{SpatICS.individual}{a vector of Spatial individual complementary specialization
    indices for each individual of the population (the overlap between
    each individual and the population UDs).}
\item{SpatICS.population}{a value of Spatial individual complementary specialization
    for the population, calculated as the average of all SpatICS individual values.}
}
\description{
This function calculates utilization distributions (UD) for both
the individuals and the whole (or rest of the) population, based on
telemetry location points, and calculates both the Spatial Individual Specialization
Index (\code{SpatIS}) and the Spatial Individual Complementary Specialization
Index (\code{SpatICS}).
The individual SpatIS is calculated as the volume of the individual UD
(or a X\% kernel density estimation area, X defined by the user) that do
not overlap with the populational UD (or a correspondent population KDE area).
The individual SpatICS is calculated as the volume of the individual UD
(or a X\% KDE area, X defined by the user) that do
not overlap with the UD of the rest of the population, after excluding that individual
(or a correspondent KDE area of the rest of the population).
The population SpatIS and SpatICS are the average of the the individual
SpatIS and SpatICS values, averaged over all individuals.
}
\examples{
library(adehabitatHR)

# raw data table
data(bats)

# run SpatIS with SpatialPoints as input

# remove individual with small sample
bats_sp <- bats_sp[bats_sp$Animal_ID != 11,]
#run
spatis(data = bats_sp, individuals.col = "Animal_ID", population.ID = NULL,
       index = "spatis")

spatis(data = bats_sp, individuals.col = "Animal_ID", population.ID = NULL,
       index = c("spatis", "SpatICS"), method = "UDOI", percent = 90)

#---
# Do NOT run, not working for SpatICS
# run SpatIS with sf as input

# remove individual with small sample
bats_sf <- bats_sf[bats_sf$Animal_ID != 11,]
#run
spatis(data = bats_sf, individuals.col = "Animal_ID", population.ID = NULL,
       index = c("spatis", "SpatICS"))

}
\references{
Kerches-Rogeri, P., Niebuhr, B.B., Muylaert, R.L, Mello, M.A.R.  Individual
specialization in the space use of frugivorous bats. Journal of Animal Ecology.
Fieberg, J. and Kochanny, C.O. (2005) Quantifying home-range overlap: the importance of
the utilization distribution. Journal of Wildlife Management, 69, 1346–1359.
}
\seealso{
\code{\link[adehabitatHR]{kernelUD}} for the estimation of utilization distributions, and
\code{\link[adehabitatHR]{kerneloverlap}} for the calculation of the overlap between utilization distributions or
areas of use.
}
\author{
Bernardo B. Niebuhr <bernardo_brandaum@yahoo.com.br> and Patricia Kerches-Rogeri
<parogeri@gmail.com>.
}
